+++ spotbugs/src/main/java/edu/umd/cs/findbugs/AbstractBugReporter.java
@@ -279,7 +279,7 @@ public abstract class AbstractBugReporter implements BugReporter {
             return;
         }
 
-        logMissingClass(classDescriptor.toDottedClassName());
+        logMissingClass(classDescriptor.getDottedClassName());
     }
 
     /**