+++ spotbugs/src/main/java/edu/umd/cs/findbugs/DiscoverSourceDirectories.java
@@ -311,7 +311,7 @@ public class DiscoverSourceDirectories {
 
             return packageName + sourceFile;
         } catch (CheckedAnalysisException e) {
-            errorLogger.logError("Could scan class " + classDesc.toDottedClassName(), e);
+            errorLogger.logError("Could scan class " + classDesc.getDottedClassName(), e);
             throw e;
         } finally {
             progress.finishClass();
@@ -331,7 +331,7 @@ public class DiscoverSourceDirectories {
                 String fullyQualifiedSourceFileName = findFullyQualifiedSourceFileName(classPath, classDesc);
                 fullyQualifiedSourceFileNameList.add(fullyQualifiedSourceFileName);
             } catch (IOException | CheckedAnalysisException e) {
-                errorLogger.logError("Couldn't scan class " + classDesc.toDottedClassName(), e);
+                errorLogger.logError("Couldn't scan class " + classDesc.getDottedClassName(), e);
             }
         }
 
@@ -400,7 +400,7 @@ public class DiscoverSourceDirectories {
 
             @Override
             public void reportMissingClass(ClassDescriptor classDescriptor) {
-                logError("Missing class: " + classDescriptor.toDottedClassName());
+                logError("Missing class: " + classDescriptor.getDottedClassName());
             }
 
             @Override