+++ CHANGELOG.md
@@ -56,6 +56,7 @@ Currently the versioning policy of this project follows [Semantic Versioning v2.
 - Rename shadowing fields ([#3221](https://github.com/spotbugs/spotbugs/pull/3221))
 - Combine catch blocks with the same body ([#3223](https://github.com/spotbugs/spotbugs/pull/3223))
 - Merge conditions of nested ifs ([#3231](https://github.com/spotbugs/spotbugs/pull/3231))
+- Use non deprecated 'getDottedClassName' instead of 'toDottedClassName'([#3251](https://github.com/spotbugs/spotbugs/pull/3251))
 
 ### Changed
 - Bump up Java version to 11